requirements.txt
langchain==0.2.10
langchain-anthropic==0.1.20
langchain-openai==0.1.17
pyyaml==6.0.1
python-dotenv==1.0.1
langsmith==0.1.93